+++ ProjectForAccountSystemRepository/src/main/java/za/ac/nwu/ac/repo/persistence/MemberRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import za.ac.nwu.ac.domain.persistence.Member;
 @Repository
-public interface MemberRepository extends JpaRepository<Member,Long>
+public interface MemberRepository extends JpaRepository<Member,Integer>
 {
     @Query(value = "SELECT"
             +"*"
